+++ src/theory/arith/fc_simplex.cpp
@@ -19,7 +19,7 @@
 #include "options/arith_options.h"
 #include "smt/smt_statistics_registry.h"
 #include "theory/arith/linear/constraint.h"
-#include "theory/arith/error_set.h"
+#include "theory/arith/linear/error_set.h"
 #include "util/statistics_stats.h"
 
 using namespace std;
