+++ src/theory/arith/linear/cut_log.h
@@ -32,7 +32,7 @@
 
 namespace cvc5::internal {
 namespace theory {
-namespace arith {
+namespace arith::linear {
 
 /** A low level vector of indexed doubles. */
 struct PrimitiveVec {
