+++ src/theory/arith/linear/error_set.cpp
@@ -16,7 +16,7 @@
  * \todo document this file
  */
 
-#include "theory/arith/error_set.h"
+#include "theory/arith/linear/error_set.h"
 
 #include "smt/smt_statistics_registry.h"
 #include "theory/arith/linear/constraint.h"
