+++ src/theory/arith/linear/error_set.h
@@ -29,7 +29,7 @@
 #include "theory/arith/linear/callbacks.h"
 #include "theory/arith/delta_rational.h"
 #include "theory/arith/linear/partial_model.h"
-#include "theory/arith/tableau_sizes.h"
+#include "theory/arith/linear/tableau_sizes.h"
 #include "util/bin_heap.h"
 #include "util/statistics_stats.h"
