+++ src/theory/arith/linear/simplex.h
@@ -67,7 +67,7 @@
 
 namespace cvc5::internal {
 namespace theory {
-namespace arith {
+namespace arith::linear {
 
 class ErrorSet;
 class LinearEqualityModule;
