+++ src/theory/arith/linear/theory_arith_private.cpp
@@ -52,7 +52,7 @@
 #include "theory/arith/arithvar.h"
 #include "theory/arith/linear/congruence_manager.h"
 #include "theory/arith/linear/constraint.h"
-#include "theory/arith/cut_log.h"
+#include "theory/arith/linear/cut_log.h"
 #include "theory/arith/delta_rational.h"
 #include "theory/arith/linear/dio_solver.h"
 #include "theory/arith/linear_equality.h"
@@ -60,7 +60,7 @@
 #include "theory/arith/nl/nonlinear_extension.h"
 #include "theory/arith/linear/normal_form.h"
 #include "theory/arith/linear/partial_model.h"
-#include "theory/arith/simplex.h"
+#include "theory/arith/linear/simplex.h"
 #include "theory/arith/theory_arith.h"
 #include "theory/ext_theory.h"
 #include "theory/quantifiers/fmf/bounded_integers.h"
