+++ src/theory/arith/linear/theory_arith_private.h
@@ -32,14 +32,14 @@
 #include "theory/arith/arith_static_learner.h"
 #include "theory/arith/arith_utilities.h"
 #include "theory/arith/arithvar.h"
-#include "theory/arith/attempt_solution_simplex.h"
+#include "theory/arith/linear/attempt_solution_simplex.h"
 #include "theory/arith/branch_and_bound.h"
 #include "theory/arith/linear/congruence_manager.h"
 #include "theory/arith/linear/constraint.h"
 #include "theory/arith/delta_rational.h"
 #include "theory/arith/linear/dio_solver.h"
-#include "theory/arith/dual_simplex.h"
-#include "theory/arith/error_set.h"
+#include "theory/arith/linear/dual_simplex.h"
+#include "theory/arith/linear/error_set.h"
 #include "theory/arith/fc_simplex.h"
 #include "theory/arith/infer_bounds.h"
 #include "theory/arith/linear_equality.h"
@@ -47,7 +47,7 @@
 #include "theory/arith/linear/normal_form.h"
 #include "theory/arith/linear/partial_model.h"
 #include "theory/arith/proof_checker.h"
-#include "theory/arith/soi_simplex.h"
+#include "theory/arith/linear/soi_simplex.h"
 #include "theory/arith/theory_arith.h"
 #include "theory/valuation.h"
 #include "util/dense_map.h"
